Split Playwright interaction audit helpers
