+++ docs/DECISIONS.md
@@ -524,3 +524,4 @@
 - Recursive subagent delegation is part of the slash-command workflow, not just the scheduler primitive. A worker may emit bounded `[[DELEGATE: name | role]]` markers, the slash runner parses them into child workers through `WorkspaceSubagentSpawnDirectiveParser`, and Activity must preserve the nested path so delegated work stays replayable and auditable from the normal thread surface.
 - Interactive SwiftUI containers must declare click-target spacing explicitly. `HStack`, `LazyHGrid`, and `LazyVGrid` clusters that contain multiple controls may not rely on SwiftUI's default spacing because that makes peer clearance an implicit local choice. The source gate now rejects both raw numeric spacing and omitted spacing for interactive clusters, while allowing one-control passive rows such as text plus a trailing Retry button.
 - Sidebar visibility is workspace chrome state, not a secondary pane. `WorkspaceChromeState` is the model-owned source of truth, `WorkspaceChromeSurface` is projected into native SwiftUI/static HTML/Playwright, and `toggle-sidebar` routes through the same typed command plan/executor path as other workspace commands. Do not add one-off sidebar booleans in renderer or harness code; new chrome visibility controls should extend this contract.
+- Playwright interaction auditing has a public façade plus focused implementation modules. Tests should keep importing from `interaction-audit-helpers.ts`, but new constants/types belong in `interaction-audit-contracts.ts`, browser-side geometry/a11y report logic belongs in `interaction-audit-report.ts`, and Playwright assertion/registry/click-probe helpers belong in `interaction-audit-targets.ts`. Keep `page.evaluate` browser code self-contained unless we intentionally introduce an injected browser-audit bundle.
